# Prepare to build docs site with strict flag for validation