--- a/.github/workflows/prepare_mkdocs.sh
+++ b/.github/workflows/prepare_mkdocs.sh
@@ -6,7 +6,7 @@
 # Install the packages with the following command:
 # pip install mkdocs mkdocs-material
 
-set -ex
+set -e
 
 # Dokka filenames like `-http-url/index.md` don't work well with MkDocs <title> tags.
 # Assign metadata to the file's first Markdown heading.
@@ -18,7 +18,9 @@ title_markdown_file() {
   echo -e "---\n$TITLE\n---\n\n$CONTENT" > "$1"
 }
 
-set +x
+echo "Prepare Misk Docs Site"
+
+echo "Add titles to Dokka Markdown files"
 for MARKDOWN_FILE in $(find docs/0.x/ -name '*.md'); do
   if grep -q "title:" $MARKDOWN_FILE; then
     continue
@@ -27,45 +29,35 @@ for MARKDOWN_FILE in $(find docs/0.x/ -name '*.md'); do
   echo $MARKDOWN_FILE
   title_markdown_file $MARKDOWN_FILE
 done
-set -x
 
-# Copy in special files that GitHub wants in the project root.
+echo "Copy in special files that GitHub wants in the project root"
 cat README.md | grep -v 'project website' > docs/index.md
 cp CHANGELOG.md docs/changelog.md
 cp CONTRIBUTING.md docs/contributing.md
 
-# Copy in Wisp docs
-#mkdir -p docs/wisp
-#cp wisp/README.md docs/wisp/readme.md
-
-# Define the source and target directories
+echo "Copy in Wisp docs"
 src_dir="wisp"
 target_dir="docs"
 
 # Use find to locate all README.md files in all subdirectories
-set +x
 find "$src_dir" -name 'README.md' -print0 | while IFS= read -r -d '' file; do
     # Create the target directory, preserving the directory structure
     mkdir -p "$target_dir/$(dirname "$file")"
     # Copy the file
     cp "$file" "$target_dir/$(dirname "$file")/index.md"
 done
-set -x
 
-# Fix docs/ relative links in special files
+echo "Fix docs/ relative links in special files"
 dir="docs"
-set +x
 find "$dir" -name '*.md' -print0 | while IFS= read -r -d '' file; do
   # Use sed to replace the links and store the result in a variable
   modified_content=$(sed -E 's|\(\./docs/(.*).md\)|(\1/)|g' "$file")
   # Echo the modified content back into the file
   echo "$modified_content" > "$file"
 done
-set -x
 
-# Fix wisp/ relative links
+echo "Fix wisp/ relative links"
 dir="docs/wisp"
-set +x
 # Use find to locate all files in the directory
 find "$dir" -type f -print0 | while IFS= read -r -d '' file; do
   # Use sed to replace the links and store the result in a variable
@@ -73,5 +65,3 @@ find "$dir" -type f -print0 | while IFS= read -r -d '' file; do
   # Echo the modified content back into the file
   echo "$modified_content" > "$file"
 done
-set -x
-