--- a/.github/workflows/prepare_mkdocs.sh
+++ b/.github/workflows/prepare_mkdocs.sh
@@ -20,6 +20,10 @@ title_markdown_file() {
 
 set +x
 for MARKDOWN_FILE in $(find docs/0.x/ -name '*.md'); do
+  if grep -q "title:" $MARKDOWN_FILE; then
+    continue
+  fi
+
   echo $MARKDOWN_FILE
   title_markdown_file $MARKDOWN_FILE
 done
@@ -39,9 +43,32 @@ src_dir="wisp"
 target_dir="docs"
 
 # Use find to locate all README.md files in all subdirectories
+set +x
 find "$src_dir" -name 'README.md' -print0 | while IFS= read -r -d '' file; do
     # Create the target directory, preserving the directory structure
     mkdir -p "$target_dir/$(dirname "$file")"
     # Copy the file
     cp "$file" "$target_dir/$(dirname "$file")/readme.md"
 done
+set -x
+
+# Fix docs/ relative links in special files
+
+dir="docs"
+set +x
+find "$dir" -name '*.md' -print0 | while IFS= read -r -d '' file; do
+  # Use sed to replace the links and store the result in a variable
+  modified_content=$(sed -E 's|\(\./docs/(.*).md\)|(\1/)|g' "$file")
+  # Echo the modified content back into the file
+  echo "$modified_content" > "$file"
+done
+set -x
+# Use find to locate all files in the directory
+#find "$dir" -type f -print0 | while IFS= read -r -d '' file; do
+#  # Use sed to replace the links and store the result in a variable
+#  modified_content=$(sed -E 's|\(\./wisp/(.*).md\)|(\1/)|g' "$file")
+#  # Echo the modified content back into the file
+#  set +x
+#  echo "$modified_content" > "$file"
+#  set -x
+#done