--- a/.gitignore
+++ b/.gitignore
@@ -26,4 +26,6 @@ hooks.gradle
 docs/0.x/*
 docs/index.md
 docs/changelog.md
+docs/contributing.md
 docs/releasing.md
+docs/wisp/readme.md