--- a/README.md
+++ b/README.md
@@ -1,5 +1,7 @@
 <img src="https://github.com/cashapp/misk/raw/master/misk.png" width="300">
 
+[<img src="https://img.shields.io/maven-central/v/com.squareup.misk/misk.svg?label=latest%20release"/>](http://search.maven.org/#search%7Cga%7C1%7Ccom.squareup.misk)
+
 * Releases
     * See most recent [public build][snap]
     * [change log][changelog]