--- a/docs/developer-guide.md
+++ b/docs/developer-guide.md
@@ -5,13 +5,13 @@ Developer Guide
 Actions are Misk's unit for an endpoint. Misk lets you make HTTP actions, and gRPC actions via
 [Wire](https://github.com/square/wire).
 
-Learn more about [Misk actions](./docs/actions.md)
+Learn more about [Misk actions](actions/)
 
 ## Interceptors
 Misk uses interceptors (middleware) to observe and potentially transform HTTP messages. The pattern
 is borrowed from [OkHttp Interceptors].
 
-Learn more about [Misk interceptors](./docs/interceptors.md)
+Learn more about [Misk interceptors](interceptors/)
 
 ## Clients
 Misk provides configurable HTTP clients built on top of [OkHttp](https://github.com/square/okhttp)
@@ -19,18 +19,18 @@ and [Retrofit](https://github.com/square/retrofit), and gRPC clients built on to
 [Wire](https://github.com/square/wire).
 
 Learn more about how to create, configure and test HTTP clients or gRPC clients within a Misk
-application [here](./docs/clients.md).
+application [here](clients/).
 
 ## Tests
 Misk provides a few ways to streamline testing with JUnit.
 
-Learn more about [Misk tests](./docs/testing.md)
+Learn more about [Misk tests](testing/)
 
 ## Modules
 Misk provides dozens of modules to facilitate the development and deployment of applications, and
 the integration with clouds and various common technologies.
 
-Learn more about [Misk modules](./docs/modules.md)
+Learn more about [Misk modules](modules/)
 
 ## Wisp
 Wisp is a collection of kotlin modules providing various features and utilities, including config,