--- a/mkdocs.yml
+++ b/mkdocs.yml
@@ -74,7 +74,6 @@ nav:
       - 'Misk Metrics': migrations/misk-metrics.md
   - 'Change Log': changelog.md
   - 'Releasing': releasing.md
-  - 'Manual Releasing': releasing-manual.md
   - 'Contributing': contributing.md
   - 'Developer Guide': developer-guide.md
   - 'Stack Overflow ⏏': https://stackoverflow.com/questions/tagged/misk?sort=active