--- a/mkdocs.yml
+++ b/mkdocs.yml
@@ -69,7 +69,7 @@ nav:
   - 'Internals':
       - 'Graceful Shutdown': internals/graceful-shutdown.md
       - 'Service Management': internals/service-management.md
-      - 'Wisp': wisp/
+      - 'Wisp': wisp/index.md
   - 'Migrations':
       - 'Misk Metrics': migrations/misk-metrics.md
   - 'Change Log': changelog.md