--- a/wisp/README.md
+++ b/wisp/README.md
@@ -1,4 +1,4 @@
-# WISP
+# Wisp
 
 Wisp is a collection of kotlin modules providing various features and utilities, including config, logging, feature
 flags and more. Modules are deliberately kept small to allow choice of the exact set needed.
@@ -9,52 +9,27 @@ flags and more. Modules are deliberately kept small to allow choice of the exact
 
 Modules that are named ending in "Testing" are for test code and should never be used or included in production code.
 
-### [AWS Environment](wisp-aws-environment/README.md)
+-  [AWS Environment](wisp-aws-environment/README.md)
+-  [Client](wisp-client/README.md)
+-  [Config](wisp-config/README.md)
+-  [Container Testing](wisp-containers-testing/README.md)
+-  [Deployment](wisp-deployment/README.md)
+-  [Deployment Testing](wisp-deployment-testing/README.md)
+-  [Feature Flags](wisp-feature/README.md)
+-  [Feature Flags Testing](wisp-feature-testing/README.md)
+-  [LaunchDarkly](wisp-launchdarkly/README.md)
+-  [Lease](wisp-lease/README.md)
+-  [Lease Testing](wisp-lease-testing/README.md)
+-  [Logging](wisp-logging/README.md)
+-  [Logging Testing](wisp-logging-testing/README.md)
+-  [Moshi](wisp-moshi/README.md)
+-  [Resource Loader](wisp-resource-loader/README.md)
+-  [Resource Loader Testing](wisp-resource-loader-testing/README.md)
+-  [Sampling](wisp-sampling/README.md)
+-  [SSL](wisp-ssl/README.md)
+-  [Tasks](wisp-task/README.md)
+-  [Time Testing](wisp-time-testing/README.md)
+-  [Token](wisp-token/README.md)
+-  [Token Testing](wisp-token-testing/README.md)
+-  [Tracing](wisp-tracing/README.md)
 
-### [Client](wisp-client/README.md)
-
-### [Config](wisp-config/README.md)
-
-### [Container Testing](wisp-containers-testing/README.md)
-
-### [Deployment](wisp-deployment/README.md)
-
-### [Deployment Testing](wisp-deployment-testing/README.md)
-
-### [Feature Flags](wisp-feature/README.md)
-
-### [Feature Flags Testing](wisp-feature-testing/README.md)
-
-### [LaunchDarkly](wisp-launchdarkly/README.md)
-
-### [Lease](wisp-lease/README.md)
-
-### [Lease Testing](wisp-lease-testing/README.md)
-
-### [Logging](wisp-logging/README.md)
-
-### [Logging Testing](wisp-logging-testing/README.md)
-
-### [Moshi](wisp-moshi/README.md)
-
-### [Resource Loader](wisp-resource-loader/README.md)
-
-### [Resource Loader Testing](wisp-resource-loader-testing/README.md)
-
-### [Sampling](wisp-sampling/README.md)
-
-### [SSL](wisp-ssl/README.md)
-
-### [Tasks](wisp-task/README.md)
-
-### [Time Testing](wisp-time-testing/README.md)
-
-### [Token](wisp-token/README.md)
-
-### [Token Testing](wisp-token-testing/README.md)
-
-### [Tracing](wisp-tracing/README.md)
-
-## Changelog
-
-See a list of changes in each release [here](CHANGELOG.md)