--- a/wisp/README.md
+++ b/wisp/README.md
@@ -1,9 +1,7 @@
 # Wisp
 
-Wisp is a collection of kotlin modules providing various features and utilities, including config, logging, feature
-flags and more. Modules are deliberately kept small to allow choice of the exact set needed.
-
-[<img src="https://img.shields.io/maven-central/v/app.cash.wisp/wisp-logging.svg?label=latest%20release"/>](http://search.maven.org/#search%7Cga%7C1%7Capp.cash.wisp)
+Wisp is a collection of Misk modules providing various features and utilities, including config, logging, feature
+flags and more, but notably implemented generally without the use of Guice Direct Injection.
 
 ## Modules