--- a/wisp/build.gradle.kts
+++ b/wisp/build.gradle.kts
@@ -1,5 +1,4 @@
 subprojects {
-
     apply(plugin = "java")
     apply(plugin = "kotlin")
     apply(plugin = "com.google.protobuf")