… bin/.python3-3.12.2.pkg
… bin/pip
… bin/pip3
… bin/pip3.12
… bin/pydoc3
… bin/pydoc3.12
… bin/python
… bin/python3
… bin/python3-config
… bin/python3.12
… bin/python3.12-config