- Start fixing errors to build docs site with strict flag for link validation etc.
- Include Wisp docs in site.
- Update changelog language.
- Add Python3 with Hermit for mkdocs build.

- Optimizes the docs build script to speed up builds.

  Old
  ---
  ```
  ./.github/workflows/prepare_mkdocs.sh  11.76s user 37.81s system 68% cpu 1:12.41 total
  ```
  
  New
  ---
  ```
  ./.github/workflows/prepare_mkdocs.sh  10.51s user 32.58s system 75% cpu 56.781 total
  ```