@adrw 2024-05-13
Subsequent PRs will fix remaining `mkdocs build --strict` issues.